broke out low level SSL tasks into separate file
