+++ patchwork/tests/test_feature_rotation.py
@@ -2,24 +2,24 @@
 import numpy as np
 import tensorflow as tf
 
-from patchwork.feature._rotation import _rotate, _build_rotation_dataset
 from patchwork.feature._rotation import _build_rotation_training_step
 
-
+def test_build_rotation_training_step():
+    inpt = tf.keras.layers.Input((None, None, 3))
+    net = tf.keras.layers.Conv2D(2,1)(inpt)
+    net = tf.keras.layers.GlobalMaxPool2D()(net)
+    net = tf.keras.layers.Dense(4, activation="softmax")(net)
+    model = tf.keras.Model(inpt, net)
+    opt = tf.keras.optimizers.SGD()
     
-def test_build_rotation_dataset(test_png_path):
-    filepaths = 10*[test_png_path]
-    ds = _build_rotation_dataset(filepaths, imshape=(32,32),
-                                     batch_size=5, 
-                                     augment={"flip_left_right":True})
-    assert isinstance(ds, tf.data.Dataset)
-    for x,y in ds:
-        x = x.numpy()
-        y = y.numpy()
-        break
+    step = _build_rotation_training_step(model, opt)
     
-    assert x.shape == (5,32,32,3)
-    assert y.shape == (5,)
-    assert y.max() < 4
-    assert y.min() >= 0
+    x = np.zeros((7,16,16,3)).astype(np.float32)
+    y = np.zeros(7).astype(np.float64)
+    
+    lossdict = step(x,y)
+    assert isinstance(lossdict, dict)
+    loss = lossdict["loss"].numpy()
+    assert loss.shape == ()
+    assert loss.dtype == np.float32
